Backend user model

+++ frontend/src/components/TaskDetails.js
@@ -1,12 +1,20 @@
 import useTasksContext from "../hooks/useTasksContext";
 import moment from 'moment';
 import {ImBin} from 'react-icons/im'
+import useAuthContext from "../hooks/useAuthContext";
 
 const TaskDetails = ({task}) => {
     const { dispatch } = useTasksContext();
+    const { user } = useAuthContext();
     const handleDelete = async () => {
+        if(!user){
+            return
+        }
         const response = await fetch(`/api/tasks/${task._id}`, {
-            method: 'DELETE'
+            method: 'DELETE',
+            headers: {
+                'Authorization': `Bearer ${user.token}`
+            }
         });
         const json = await response.json();
         if(response.ok){
